api/train/requirements.txt
pandas>=1.5.3
numpy>=1.23.5
scikit-learn>=1.2.2
wandb>=0.15.8
joblib>=1.2.0
python-dotenv>=0.21.1